勉强可以达到通过移动节点配置节点位置这个功能 移动节点并点击"确认变更"按键之后，节点会有轻微的移动

+++ WpfApp1/MainWindow.xaml.cs
@@ -175,8 +175,8 @@ namespace WpfApp1
                 int index_tempj = Convert.ToInt16(temp_DataRow_tt[i][0]) - 1;
                 Ellipse_Array_tt[index_tempj].Visibility = Visibility.Visible;
                 change_XY_rectangle(Ellipse_Array_tt[index_tempj], 
-                                    Convert.ToInt16(temp_DataRow_tt[i][1]), 
-                                    Convert.ToInt16(temp_DataRow_tt[i][2]));
+                                    Convert.ToDouble(temp_DataRow_tt[i][1]), 
+                                    Convert.ToDouble(temp_DataRow_tt[i][2]));
             }
         }
